requirements.txt
fastapi
uvicorn
opencv-python
torch
mediapipe
ultralytics
google-generativeai
python-dotenv